requirements.txt
Django>=3.0.4
django-filter>=2.4.0